AG-7768 - Edits, Deletes, FillHandle and Clipboard should fire `onCellEditRequest`
