+++ community-modules/core/src/ts/cellNavigationService.ts
@@ -2,10 +2,10 @@ import { Autowired, Bean } from "./context/context";
 import { BeanStub } from "./context/beanStub";
 import { ColumnModel } from "./columns/columnModel";
 import { IRowModel } from "./interfaces/iRowModel";
-import { CellPosition } from "./entities/cellPosition";
+import { CellPosition } from "./entities/cellPositionUtils";
 import { RowNode } from "./entities/rowNode";
 import { Column } from "./entities/column";
-import { RowPosition } from "./entities/rowPosition";
+import { RowPosition } from "./entities/rowPositionUtils";
 import { PinnedRowModel } from "./pinnedRowModel/pinnedRowModel";
 import { missing } from "./utils/generic";
 import { last } from "./utils/array";
