+++ community-modules/core/src/ts/focusService.ts
@@ -3,10 +3,10 @@ import { BeanStub } from "./context/beanStub";
 import { Column } from "./entities/column";
 import { CellFocusedParams, CellFocusedEvent, Events } from "./events";
 import { ColumnModel } from "./columns/columnModel";
-import { CellPosition } from "./entities/cellPosition";
+import { CellPosition } from "./entities/cellPositionUtils";
 import { RowNode } from "./entities/rowNode";
 import { HeaderPosition } from "./headerRendering/common/headerPosition";
-import { RowPositionUtils } from "./entities/rowPosition";
+import { RowPositionUtils } from "./entities/rowPositionUtils";
 import { IRangeService } from "./interfaces/IRangeService";
 import { RowRenderer } from "./rendering/rowRenderer";
 import { HeaderNavigationService } from "./headerRendering/common/headerNavigationService";
