+++ community-modules/core/src/ts/grid.ts
@@ -67,8 +67,8 @@ import { AgInputNumberField } from "./widgets/agInputNumberField";
 import { AgInputRange } from "./widgets/agInputRange";
 import { AgSelect } from "./widgets/agSelect";
 import { AgToggleButton } from "./widgets/agToggleButton";
-import { RowPositionUtils } from "./entities/rowPosition";
-import { CellPositionUtils } from "./entities/cellPosition";
+import { RowPositionUtils } from "./entities/rowPositionUtils";
+import { CellPositionUtils } from "./entities/cellPositionUtils";
 import { PinnedRowModel } from "./pinnedRowModel/pinnedRowModel";
 import { ModuleRegistry } from "./modules/moduleRegistry";
 import { ModuleNames } from "./modules/moduleNames";
