+++ community-modules/core/src/ts/rendering/beans.ts
@@ -24,8 +24,8 @@ import { SortController } from "../sortController";
 import { FilterManager } from "../filter/filterManager";
 import { RowContainerHeightService } from "./rowContainerHeightService";
 import { IFrameworkOverrides } from "../interfaces/iFrameworkOverrides";
-import { CellPositionUtils } from "../entities/cellPosition";
-import { RowPositionUtils } from "../entities/rowPosition";
+import { CellPositionUtils } from "../entities/cellPositionUtils";
+import { RowPositionUtils } from "../entities/rowPositionUtils";
 import { SelectionService } from "../selectionService";
 import { RowCssClassCalculator } from "./row/rowCssClassCalculator";
 import { IRowModel } from "../interfaces/iRowModel";
