+++ community-modules/core/src/ts/undoRedo/undoRedoService.ts
@@ -5,11 +5,11 @@ import { FocusService } from "../focusService";
 import { IRowModel } from "../interfaces/iRowModel";
 import { PinnedRowModel } from "../pinnedRowModel/pinnedRowModel";
 import { CellValueChange, RangeUndoRedoAction, LastFocusedCell, UndoRedoAction, UndoRedoStack } from "./undoRedoStack";
-import { RowPosition, RowPositionUtils } from "../entities/rowPosition";
+import { RowPosition, RowPositionUtils } from "../entities/rowPositionUtils";
 import { RowNode } from "../entities/rowNode";
 import { CellRange, CellRangeParams, IRangeService } from "../interfaces/IRangeService";
 import { BeanStub } from "../context/beanStub";
-import { CellPosition, CellPositionUtils } from "../entities/cellPosition";
+import { CellPosition, CellPositionUtils } from "../entities/cellPositionUtils";
 import { Column } from '../entities/column';
 import { ColumnModel } from "../columns/columnModel";
 import { CtrlsService } from "../ctrlsService";
