+++ grid-packages/ag-grid-docs/documentation/doc-pages/grid-api/grid-options.AUTO.json
@@ -1581,7 +1581,7 @@
     "meta": { "isEvent": true, "name": "onCellValueChanged" }
   },
   "onCellEditRequest": {
-    "description": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+    "description": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
     "type": {
       "arguments": { "event": "CellEditRequestEvent<TData>" },
       "returnType": "void"
@@ -1589,7 +1589,7 @@
     "meta": { "isEvent": true, "name": "onCellEditRequest" }
   },
   "cellEditRequest": {
-    "description": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+    "description": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
     "type": {
       "arguments": { "event": "CellEditRequestEvent<TData>" },
       "returnType": "void"
