+++ grid-packages/ag-grid-docs/documentation/doc-pages/grid-api/interfaces.AUTO.json
@@ -2201,7 +2201,7 @@
       "onColumnEverythingChanged?(event: ColumnEverythingChangedEvent<TData>)": "/** Shotgun - gets called when either a) new columns are set or b) `columnApi.applyColumnState()` is used, so everything has changed. */",
       "onComponentStateChanged?(event: ComponentStateChangedEvent<TData>)": "/** Only used by Angular, React and VueJS AG Grid components (not used if doing plain JavaScript).\n     * If the grid receives changes due to bound properties, this event fires after the grid has finished processing the change.\n     */",
       "onCellValueChanged?(event: CellValueChangedEvent<TData>)": "/** Value has changed after editing (this event will not fire if editing was cancelled, eg ESC was pressed) or\n     *  if cell value has changed as a result of paste operation.\n    */",
-      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
       "onRowValueChanged?(event: RowValueChangedEvent<TData>)": "/** A cell's value within a row has changed. This event corresponds to Full Row Editing only. */",
       "onCellEditingStarted?(event: CellEditingStartedEvent<TData>)": "/** Editing a cell has started. */",
       "onCellEditingStopped?(event: CellEditingStoppedEvent<TData>)": "/** Editing a cell has stopped. */",
@@ -12191,7 +12191,7 @@
       "onColumnEverythingChanged?(event: ColumnEverythingChangedEvent<TData>)": "/** Shotgun - gets called when either a) new columns are set or b) `columnApi.applyColumnState()` is used, so everything has changed. */",
       "onComponentStateChanged?(event: ComponentStateChangedEvent<TData>)": "/** Only used by Angular, React and VueJS AG Grid components (not used if doing plain JavaScript).\n     * If the grid receives changes due to bound properties, this event fires after the grid has finished processing the change.\n     */",
       "onCellValueChanged?(event: CellValueChangedEvent<TData>)": "/** Value has changed after editing (this event will not fire if editing was cancelled, eg ESC was pressed) or\n     *  if cell value has changed as a result of paste operation.\n    */",
-      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
       "onRowValueChanged?(event: RowValueChangedEvent<TData>)": "/** A cell's value within a row has changed. This event corresponds to Full Row Editing only. */",
       "onCellEditingStarted?(event: CellEditingStartedEvent<TData>)": "/** Editing a cell has started. */",
       "onCellEditingStopped?(event: CellEditingStoppedEvent<TData>)": "/** Editing a cell has stopped. */",
@@ -12936,7 +12936,7 @@
       "onColumnEverythingChanged?(event: ColumnEverythingChangedEvent<TData>)": "/** Shotgun - gets called when either a) new columns are set or b) `columnApi.applyColumnState()` is used, so everything has changed. */",
       "onComponentStateChanged?(event: ComponentStateChangedEvent<TData>)": "/** Only used by Angular, React and VueJS AG Grid components (not used if doing plain JavaScript).\n     * If the grid receives changes due to bound properties, this event fires after the grid has finished processing the change.\n     */",
       "onCellValueChanged?(event: CellValueChangedEvent<TData>)": "/** Value has changed after editing (this event will not fire if editing was cancelled, eg ESC was pressed) or\n     *  if cell value has changed as a result of paste operation.\n    */",
-      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
       "onRowValueChanged?(event: RowValueChangedEvent<TData>)": "/** A cell's value within a row has changed. This event corresponds to Full Row Editing only. */",
       "onCellEditingStarted?(event: CellEditingStartedEvent<TData>)": "/** Editing a cell has started. */",
       "onCellEditingStopped?(event: CellEditingStoppedEvent<TData>)": "/** Editing a cell has stopped. */",
@@ -13683,7 +13683,7 @@
       "onColumnEverythingChanged?(event: ColumnEverythingChangedEvent<TData>)": "/** Shotgun - gets called when either a) new columns are set or b) `columnApi.applyColumnState()` is used, so everything has changed. */",
       "onComponentStateChanged?(event: ComponentStateChangedEvent<TData>)": "/** Only used by Angular, React and VueJS AG Grid components (not used if doing plain JavaScript).\n     * If the grid receives changes due to bound properties, this event fires after the grid has finished processing the change.\n     */",
       "onCellValueChanged?(event: CellValueChangedEvent<TData>)": "/** Value has changed after editing (this event will not fire if editing was cancelled, eg ESC was pressed) or\n     *  if cell value has changed as a result of paste operation.\n    */",
-      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when doing `readOnlyEdit=true`. */",
+      "onCellEditRequest?(event: CellEditRequestEvent<TData>)": "/** Value has changed after editing. Only fires when `readOnlyEdit=true`. */",
       "onRowValueChanged?(event: RowValueChangedEvent<TData>)": "/** A cell's value within a row has changed. This event corresponds to Full Row Editing only. */",
       "onCellEditingStarted?(event: CellEditingStartedEvent<TData>)": "/** Editing a cell has started. */",
       "onCellEditingStopped?(event: CellEditingStoppedEvent<TData>)": "/** Editing a cell has stopped. */",
